Added scripts to compute errors and also convergence. As expected, the Euler solver conveeges with dx^2
One caution when doing such tests: In G2 the last output file may not
be written exactly at tEnd but my be offset by as much as dt. Hence it
is important to compute convergence etc by looking at the actual time
recorded in the BP file and not assuming it is tEnd. Misleading
results may otherwise result.